move code torecentlyAdded.ts since logic is the same with some label changes
